+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v2.5.2
+- (DEV ONLY) now uses `matjson` instead of manual parsing
+- ui bug fix
+- another UI bug (spinning circle is on top right sorry it'll be fixed soon)
+- i forgot what else...
+
 ## v2.5.1
 - blackscreen toggle works for mobile now by adding a button